Copy immutable data in CommentForm
